add auth

+++ app/Http/Controllers/TasklistsController.php
@@ -17,12 +17,21 @@ class TasklistsController extends Controller
      */
     public function index()
     {
-        $tasklists = Tasklist::all();
+        $tasklist = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasklists = $user->tasklists()->orderBy('created_at', 'desc')->paginate(10);
 
-        return view('tasklists.index', [
-            'tasklists' => $tasklists,
-        ]);
+            $tasklist = [
+                'user' => $user,
+                'tasklists' => $tasklists,
+            ];
+            return view('tasklists.index', $tasklist);
+        }else {
+            return view('welcome');
+        }
     }
+
 
     /**
      * Show the form for creating a new resource.
@@ -55,6 +64,7 @@ class TasklistsController extends Controller
          $tasklist = new Tasklist;
          $tasklist->status = $request->status;    // add
         $tasklist->content = $request->content;
+        $tasklist ->user_id = \Auth::user()->id;
         $tasklist->save();
 
         return redirect('/');
@@ -85,9 +95,22 @@ class TasklistsController extends Controller
     {
           $tasklist = Tasklist::find($id);
 
-        return view('tasklists.edit', [
+       if (!empty($tasklist->user_id)) {
+            if (\Auth::user()->id === $tasklist->user_id) {
+           return view('tasklists.edit', [
             'tasklist' => $tasklist,
-        ]);
+            ]);
+             }
+            else {
+                return redirect('/');
+            }
+        }
+        
+        else {
+            return redirect('/');
+        }
+        
+        
     }
 
     /**
@@ -106,6 +129,7 @@ class TasklistsController extends Controller
                 $tasklist = Tasklist::find($id);
                         $tasklist->status = $request->status;
                         $tasklist->content = $request->content;
+                         $tasklist ->user_id = \Auth::user()->id;
         $tasklist->save();
 
         return redirect('/');
@@ -117,10 +141,14 @@ class TasklistsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+ public function destroy($id)
     {
         $tasklist = Tasklist::find($id);
-        $tasklist->delete();
+
+        if (\Auth::user()->id === $tasklist->user_id) {
+            $tasklist->delete();
+        }
+    
 
         return redirect('/');
     }
